% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{uw_favicon}
\alias{uw_favicon}
\title{Add UW-Madison favicon to Shiny UI}
\usage{
uw_favicon()
}
\value{
A \code{tags$head()} element containing favicon link tags
}
\description{
Creates HTML head tags to include the UW-Madison favicon. This function
is called automatically by \code{uw_page_navbar()} but can be used
independently in custom page layouts.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    uw_favicon(),
    h1("My UW App")
  )
}
}
