% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{uw_nav_panel}
\alias{uw_nav_panel}
\title{Create a standard UW content panel}
\usage{
uw_nav_panel(title, ...)
}
\arguments{
\item{title}{Panel title}

\item{...}{Content to include in the panel}
}
\value{
A nav_panel element with UW styling
}
\description{
Create a standard UW content panel
}
\examples{
if (interactive()) {
  uw_nav_panel("Home", shiny::h1("Welcome"))
}
}
